+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Enhancements
 
+* Add `A.Vector.map_reduce/3`
 * Add `A.Vector.scan/2` and `A.Vector.scan/3`
 * `A.Vector.with_index/2` accepts a function parameter